adicionando exceptions

+++ src/main/java/project/ecommerce/service/ProdutoService.java
@@ -5,6 +5,7 @@ import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import project.ecommerce.exception.ResourceNotFoundException;
 import project.ecommerce.model.Produto;
 import project.ecommerce.repository.ProdutoRepository;
 
@@ -23,17 +24,26 @@ public class ProdutoService {
     }
 
     public Produto buscarPorId(Integer id) {
-        return produtoRepository.findById(id).get();
+        if (produtoRepository.existsById(id)) {
+            return produtoRepository.findById(id).get();
+        } else {
+            throw new ResourceNotFoundException("Produto com ID " + id + " não encontrado.");
+        }
     }
 
     public void deletarPorId(Integer id) {
-        produtoRepository.deleteById(id);
+        if (produtoRepository.existsById(id)) {
+            produtoRepository.deleteById(id);
+        } else {
+            throw new ResourceNotFoundException("Produto com ID " + id + " não encontrado.");
+        }
     }
 
     public Produto atualizarProduto(Produto produtoAtualizado) {
         if (produtoRepository.existsById(produtoAtualizado.getId())) {
             return produtoRepository.save(produtoAtualizado);     
+        } else {
+            throw new ResourceNotFoundException("Produto com ID " + produtoAtualizado.getId() + " não encontrado.");
         }
-        return null;
     }
 }
